backend/requirements.txt
django==2.2.19
django-extra-fields==3.0.2
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
Pillow==8.3.1
psycopg2-binary==2.8.6